refactor: webで読み込み可能に
