+++ packages/web/package.json
@@ -1,0 +1,62 @@
+{
+  "name": "web",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc -b && vite build",
+    "check:lint": "biome lint ./src",
+    "check:format": "biome format ./src",
+    "check": "run-s check:*",
+    "fix": "biome check --write ./src",
+    "preview": "vite preview",
+    "test:e2e": "playwright test",
+    "test:browser-mode": "vitest run --config=vitest.browser.config.ts",
+    "test:node": "vitest run",
+    "test": "pnpm test:node && pnpm test:browser-mode && pnpm test:e2e"
+  },
+  "dependencies": {
+    "@radix-ui/react-dialog": "^1.1.15",
+    "@radix-ui/react-popover": "^1.1.15",
+    "@radix-ui/react-slot": "^1.2.3",
+    "@tailwindcss/vite": "^4.1.11",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "lucide-react": "^0.539.0",
+    "next-themes": "^0.4.6",
+    "prismjs": "^1.30.0",
+    "prosemirror-markdown": "^1.13.2",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-ga4": "^2.1.0",
+    "sonner": "^2.0.7",
+    "tailwind-merge": "^3.3.1",
+    "tailwindcss": "^4.1.11",
+    "zenn-content-css": "^0.2.1",
+    "zenn-markdown-html": "workspace:zenn-markdown-html",
+    "zenn-wysiwyg-editor": "workspace:zenn-wysiwyg-editor"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "2.2.0",
+    "@playwright/test": "^1.55.0",
+    "@testing-library/dom": "^10.4.1",
+    "@testing-library/react": "^16.3.0",
+    "@types/node": "^24.2.1",
+    "@types/prismjs": "^1.26.5",
+    "@types/react": "^19.1.10",
+    "@types/react-dom": "^19.1.7",
+    "@vitejs/plugin-react": "^5.0.0",
+    "@vitest/browser": "^3.2.4",
+    "babel-plugin-prismjs": "^2.1.0",
+    "globals": "^16.3.0",
+    "happy-dom": "^18.0.1",
+    "npm-run-all": "^4.1.5",
+    "playwright": "^1.55.0",
+    "tw-animate-css": "^1.3.6",
+    "typescript": "~5.8.3",
+    "vite": "^7.1.2",
+    "vitest": "^3.2.4",
+    "vitest-browser-react": "^1.0.1"
+  }
+}
