+++ packages/web/src/assets/sikotuko.jpeg
