+++ packages/web/src/components/ui/dialog.tsx
@@ -1,7 +1,7 @@
 import * as DialogPrimitive from "@radix-ui/react-dialog";
 import { XIcon } from "lucide-react";
-import { cn } from "packages/web/lib/utils";
 import type * as React from "react";
+import { cn } from "../../lib/utils";
 
 function Dialog({
   ...props
