+++ packages/web/src/components/ui/input.tsx
@@ -1,5 +1,5 @@
-import { cn } from "packages/web/lib/utils";
 import type * as React from "react";
+import { cn } from "../../lib/utils";
 
 function Input({ className, type, ...props }: React.ComponentProps<"input">) {
   return (
