+++ packages/web/src/components/ui/popover.tsx
@@ -1,6 +1,6 @@
 import * as PopoverPrimitive from "@radix-ui/react-popover";
-import { cn } from "packages/web/lib/utils";
 import type * as React from "react";
+import { cn } from "../../lib/utils";
 
 function Popover({
   ...props
