+++ packages/web/src/components/ui/textarea.tsx
@@ -1,5 +1,5 @@
-import { cn } from "packages/web/lib/utils";
 import type * as React from "react";
+import { cn } from "../../lib/utils";
 
 function Textarea({ className, ...props }: React.ComponentProps<"textarea">) {
   return (
