+++ packages/web/src/hooks/use-page-tracking.ts
