+++ packages/zenn-wysiwyg-editor/package.json
@@ -47,5 +47,9 @@
   },
   "devDependencies": {
     "tsup": "^8.5.0"
+  },
+  "peerDependencies": {
+    "react": "^17.0.0 || ^18.0.0 || ^19.0.0",
+    "react-dom": "^17.0.0 || ^18.0.0 || ^19.0.0"
   }
 }
