+++ packages/zenn-wysiwyg-editor/src/index.ts
@@ -1,1 +1,3 @@
-export * from "./useZennEditor";
+export { EditorContent } from "@tiptap/react";
+export { Toaster } from "sonner";
+export { useZennEditor } from "./useZennEditor";
